change API to enable no-std: pass OKM slice *in*, rather than returning Vec
This makes the caller responsible for allocating the memory, so they can do
it on the stack if they want. We get the desired length by asking the slice
how long it is.

The deprecated derive() function retains its old API, and allocates a vector.
To make this properly no-std, we'll need to gate that function on a "use-std"
feature, or remove it.